+++ examples/sha256/table16/compression/compression_gates.rs
@@ -47,21 +47,21 @@ impl<F: FieldExt> CompressionGate<F> {
         let range_check_tag_b = Gate::range_check(tag_b, 0, 2);
         let range_check_tag_d = Gate::range_check(tag_d, 0, 1);
         let dense_check = a
-            + b * F::from_u64(1 << 2)
-            + c_lo * F::from_u64(1 << 13)
-            + c_mid * F::from_u64(1 << 16)
-            + c_hi * F::from_u64(1 << 19)
-            + d * F::from_u64(1 << 22)
+            + b * F::from(1 << 2)
+            + c_lo * F::from(1 << 13)
+            + c_mid * F::from(1 << 16)
+            + c_hi * F::from(1 << 19)
+            + d * F::from(1 << 22)
             + word_lo * (-F::one())
-            + word_hi * F::from_u64(1 << 16) * (-F::one());
+            + word_hi * F::from(1 << 16) * (-F::one());
         let spread_check = spread_a
-            + spread_b * F::from_u64(1 << 4)
-            + spread_c_lo * F::from_u64(1 << 26)
-            + spread_c_mid * F::from_u64(1 << 32)
-            + spread_c_hi * F::from_u64(1 << 38)
-            + spread_d * F::from_u64(1 << 44)
+            + spread_b * F::from(1 << 4)
+            + spread_c_lo * F::from(1 << 26)
+            + spread_c_mid * F::from(1 << 32)
+            + spread_c_hi * F::from(1 << 38)
+            + spread_d * F::from(1 << 44)
             + spread_word_lo * (-F::one())
-            + spread_word_hi * F::from_u64(1 << 32) * (-F::one());
+            + spread_word_hi * F::from(1 << 32) * (-F::one());
 
         check_spread_and_range
             .chain(Some(("range_check_tag_b", range_check_tag_b)))
@@ -112,21 +112,21 @@ impl<F: FieldExt> CompressionGate<F> {
         let range_check_tag_c = Gate::range_check(tag_c, 0, 4);
         let range_check_tag_d = Gate::range_check(tag_d, 0, 0);
         let dense_check = a_lo
-            + a_hi * F::from_u64(1 << 3)
-            + b_lo * F::from_u64(1 << 6)
-            + b_hi * F::from_u64(1 << 8)
-            + c * F::from_u64(1 << 11)
-            + d * F::from_u64(1 << 25)
+            + a_hi * F::from(1 << 3)
+            + b_lo * F::from(1 << 6)
+            + b_hi * F::from(1 << 8)
+            + c * F::from(1 << 11)
+            + d * F::from(1 << 25)
             + word_lo * (-F::one())
-            + word_hi * F::from_u64(1 << 16) * (-F::one());
+            + word_hi * F::from(1 << 16) * (-F::one());
         let spread_check = spread_a_lo
-            + spread_a_hi * F::from_u64(1 << 6)
-            + spread_b_lo * F::from_u64(1 << 12)
-            + spread_b_hi * F::from_u64(1 << 16)
-            + spread_c * F::from_u64(1 << 22)
-            + spread_d * F::from_u64(1 << 50)
+            + spread_a_hi * F::from(1 << 6)
+            + spread_b_lo * F::from(1 << 12)
+            + spread_b_hi * F::from(1 << 16)
+            + spread_c * F::from(1 << 22)
+            + spread_d * F::from(1 << 50)
             + spread_word_lo * (-F::one())
-            + spread_word_hi * F::from_u64(1 << 32) * (-F::one());
+            + spread_word_hi * F::from(1 << 32) * (-F::one());
 
         check_spread_and_range
             .chain(Some(("range_check_tag_c", range_check_tag_c)))
@@ -153,26 +153,26 @@ impl<F: FieldExt> CompressionGate<F> {
         spread_d: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let spread_witness = spread_r0_even
-            + spread_r0_odd * F::from_u64(2)
-            + (spread_r1_even + spread_r1_odd * F::from_u64(2)) * F::from_u64(1 << 32);
+            + spread_r0_odd * F::from(2)
+            + (spread_r1_even + spread_r1_odd * F::from(2)) * F::from(1 << 32);
         let xor_0 = spread_b.clone()
-            + spread_c_lo.clone() * F::from_u64(1 << 22)
-            + spread_c_mid.clone() * F::from_u64(1 << 28)
-            + spread_c_hi.clone() * F::from_u64(1 << 34)
-            + spread_d.clone() * F::from_u64(1 << 40)
-            + spread_a.clone() * F::from_u64(1 << 60);
+            + spread_c_lo.clone() * F::from(1 << 22)
+            + spread_c_mid.clone() * F::from(1 << 28)
+            + spread_c_hi.clone() * F::from(1 << 34)
+            + spread_d.clone() * F::from(1 << 40)
+            + spread_a.clone() * F::from(1 << 60);
         let xor_1 = spread_c_lo.clone()
-            + spread_c_mid.clone() * F::from_u64(1 << 6)
-            + spread_c_hi.clone() * F::from_u64(1 << 12)
-            + spread_d.clone() * F::from_u64(1 << 18)
-            + spread_a.clone() * F::from_u64(1 << 38)
-            + spread_b.clone() * F::from_u64(1 << 42);
+            + spread_c_mid.clone() * F::from(1 << 6)
+            + spread_c_hi.clone() * F::from(1 << 12)
+            + spread_d.clone() * F::from(1 << 18)
+            + spread_a.clone() * F::from(1 << 38)
+            + spread_b.clone() * F::from(1 << 42);
         let xor_2 = spread_d
-            + spread_a * F::from_u64(1 << 20)
-            + spread_b * F::from_u64(1 << 24)
-            + spread_c_lo * F::from_u64(1 << 46)
-            + spread_c_mid * F::from_u64(1 << 52)
-            + spread_c_hi * F::from_u64(1 << 58);
+            + spread_a * F::from(1 << 20)
+            + spread_b * F::from(1 << 24)
+            + spread_c_lo * F::from(1 << 46)
+            + spread_c_mid * F::from(1 << 52)
+            + spread_c_hi * F::from(1 << 58);
         let xor = xor_0 + xor_1 + xor_2;
         let check = spread_witness + (xor * -F::one());
 
@@ -196,27 +196,27 @@ impl<F: FieldExt> CompressionGate<F> {
         spread_d: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let spread_witness = spread_r0_even
-            + spread_r0_odd * F::from_u64(2)
-            + (spread_r1_even + spread_r1_odd * F::from_u64(2)) * F::from_u64(1 << 32);
+            + spread_r0_odd * F::from(2)
+            + (spread_r1_even + spread_r1_odd * F::from(2)) * F::from(1 << 32);
 
         let xor_0 = spread_b_lo.clone()
-            + spread_b_hi.clone() * F::from_u64(1 << 4)
-            + spread_c.clone() * F::from_u64(1 << 10)
-            + spread_d.clone() * F::from_u64(1 << 38)
-            + spread_a_lo.clone() * F::from_u64(1 << 52)
-            + spread_a_hi.clone() * F::from_u64(1 << 58);
+            + spread_b_hi.clone() * F::from(1 << 4)
+            + spread_c.clone() * F::from(1 << 10)
+            + spread_d.clone() * F::from(1 << 38)
+            + spread_a_lo.clone() * F::from(1 << 52)
+            + spread_a_hi.clone() * F::from(1 << 58);
         let xor_1 = spread_c.clone()
-            + spread_d.clone() * F::from_u64(1 << 28)
-            + spread_a_lo.clone() * F::from_u64(1 << 42)
-            + spread_a_hi.clone() * F::from_u64(1 << 48)
-            + spread_b_lo.clone() * F::from_u64(1 << 54)
-            + spread_b_hi.clone() * F::from_u64(1 << 58);
+            + spread_d.clone() * F::from(1 << 28)
+            + spread_a_lo.clone() * F::from(1 << 42)
+            + spread_a_hi.clone() * F::from(1 << 48)
+            + spread_b_lo.clone() * F::from(1 << 54)
+            + spread_b_hi.clone() * F::from(1 << 58);
         let xor_2 = spread_d
-            + spread_a_lo * F::from_u64(1 << 14)
-            + spread_a_hi * F::from_u64(1 << 20)
-            + spread_b_lo * F::from_u64(1 << 26)
-            + spread_b_hi * F::from_u64(1 << 30)
-            + spread_c * F::from_u64(1 << 36);
+            + spread_a_lo * F::from(1 << 14)
+            + spread_a_hi * F::from(1 << 20)
+            + spread_b_lo * F::from(1 << 26)
+            + spread_b_hi * F::from(1 << 30)
+            + spread_c * F::from(1 << 36);
         let xor = xor_0 + xor_1 + xor_2;
         let check = spread_witness + (xor * -F::one());
 
@@ -238,11 +238,11 @@ impl<F: FieldExt> CompressionGate<F> {
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let lhs_lo = spread_e_lo + spread_f_lo;
         let lhs_hi = spread_e_hi + spread_f_hi;
-        let lhs = lhs_lo + lhs_hi * F::from_u64(1 << 32);
+        let lhs = lhs_lo + lhs_hi * F::from(1 << 32);
 
-        let rhs_even = spread_p0_even + spread_p1_even * F::from_u64(1 << 32);
-        let rhs_odd = spread_p0_odd + spread_p1_odd * F::from_u64(1 << 32);
-        let rhs = rhs_even + rhs_odd * F::from_u64(2);
+        let rhs_even = spread_p0_even + spread_p1_even * F::from(1 << 32);
+        let rhs_odd = spread_p0_odd + spread_p1_odd * F::from(1 << 32);
+        let rhs = rhs_even + rhs_odd * F::from(2);
 
         let check = lhs + rhs * -F::one();
 
@@ -265,7 +265,7 @@ impl<F: FieldExt> CompressionGate<F> {
         spread_g_hi: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let neg_check = {
-            let evens = Self::ones() * F::from_u64(MASK_EVEN_32 as u64);
+            let evens = Self::ones() * F::from(MASK_EVEN_32 as u64);
             // evens - spread_e_lo = spread_e_neg_lo
             let lo_check = spread_e_neg_lo.clone() + spread_e_lo + (evens.clone() * (-F::one()));
             // evens - spread_e_hi = spread_e_neg_hi
@@ -278,11 +278,11 @@ impl<F: FieldExt> CompressionGate<F> {
 
         let lhs_lo = spread_e_neg_lo + spread_g_lo;
         let lhs_hi = spread_e_neg_hi + spread_g_hi;
-        let lhs = lhs_lo + lhs_hi * F::from_u64(1 << 32);
+        let lhs = lhs_lo + lhs_hi * F::from(1 << 32);
 
-        let rhs_even = spread_q0_even + spread_q1_even * F::from_u64(1 << 32);
-        let rhs_odd = spread_q0_odd + spread_q1_odd * F::from_u64(1 << 32);
-        let rhs = rhs_even + rhs_odd * F::from_u64(2);
+        let rhs_even = spread_q0_even + spread_q1_even * F::from(1 << 32);
+        let rhs_odd = spread_q0_odd + spread_q1_odd * F::from(1 << 32);
+        let rhs = rhs_even + rhs_odd * F::from(2);
 
         neg_check
             .chain(Some(("s_ch_neg", lhs - rhs)))
@@ -304,13 +304,13 @@ impl<F: FieldExt> CompressionGate<F> {
         spread_c_lo: Expression<F>,
         spread_c_hi: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
-        let maj_even = spread_m_0_even + spread_m_1_even * F::from_u64(1 << 32);
-        let maj_odd = spread_m_0_odd + spread_m_1_odd * F::from_u64(1 << 32);
-        let maj = maj_even + maj_odd * F::from_u64(2);
+        let maj_even = spread_m_0_even + spread_m_1_even * F::from(1 << 32);
+        let maj_odd = spread_m_0_odd + spread_m_1_odd * F::from(1 << 32);
+        let maj = maj_even + maj_odd * F::from(2);
 
-        let a = spread_a_lo + spread_a_hi * F::from_u64(1 << 32);
-        let b = spread_b_lo + spread_b_hi * F::from_u64(1 << 32);
-        let c = spread_c_lo + spread_c_hi * F::from_u64(1 << 32);
+        let a = spread_a_lo + spread_a_hi * F::from(1 << 32);
+        let b = spread_b_lo + spread_b_hi * F::from(1 << 32);
+        let c = spread_c_lo + spread_c_hi * F::from(1 << 32);
         let sum = a + b + c;
 
         std::iter::empty().chain(Some(("maj", s_maj * (sum - maj))))
@@ -339,10 +339,10 @@ impl<F: FieldExt> CompressionGate<F> {
         let lo = h_lo + ch_lo + ch_neg_lo + sigma_e_lo + k_lo + w_lo;
         let hi = h_hi + ch_hi + ch_neg_hi + sigma_e_hi + k_hi + w_hi;
 
-        let sum = lo + hi * F::from_u64(1 << 16);
-        let h_prime = h_prime_lo + h_prime_hi * F::from_u64(1 << 16);
+        let sum = lo + hi * F::from(1 << 16);
+        let h_prime = h_prime_lo + h_prime_hi * F::from(1 << 16);
 
-        let check = sum - (h_prime_carry * F::from_u64(1 << 32)) - h_prime;
+        let check = sum - (h_prime_carry * F::from(1 << 32)) - h_prime;
 
         std::iter::empty().chain(Some(("s_h_prime", s_h_prime * check)))
     }
@@ -363,10 +363,10 @@ impl<F: FieldExt> CompressionGate<F> {
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let lo = sigma_a_lo + maj_abc_lo + h_prime_lo;
         let hi = sigma_a_hi + maj_abc_hi + h_prime_hi;
-        let sum = lo + hi * F::from_u64(1 << 16);
-        let a_new = a_new_lo + a_new_hi * F::from_u64(1 << 16);
+        let sum = lo + hi * F::from(1 << 16);
+        let a_new = a_new_lo + a_new_hi * F::from(1 << 16);
 
-        let check = sum - (a_new_carry * F::from_u64(1 << 32)) - a_new;
+        let check = sum - (a_new_carry * F::from(1 << 32)) - a_new;
 
         std::iter::empty().chain(Some(("s_a_new", s_a_new * check)))
     }
@@ -385,10 +385,10 @@ impl<F: FieldExt> CompressionGate<F> {
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let lo = h_prime_lo + d_lo;
         let hi = h_prime_hi + d_hi;
-        let sum = lo + hi * F::from_u64(1 << 16);
-        let e_new = e_new_lo + e_new_hi * F::from_u64(1 << 16);
+        let sum = lo + hi * F::from(1 << 16);
+        let e_new = e_new_lo + e_new_hi * F::from(1 << 16);
 
-        let check = sum - (e_new_carry * F::from_u64(1 << 32)) - e_new;
+        let check = sum - (e_new_carry * F::from(1 << 32)) - e_new;
 
         std::iter::empty().chain(Some(("s_e_new", s_e_new * check)))
     }
@@ -411,7 +411,7 @@ impl<F: FieldExt> CompressionGate<F> {
         word_3: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let check_lo_hi = |lo: Expression<F>, hi: Expression<F>, word: Expression<F>| {
-            lo + hi * F::from_u64(1 << 16) - word
+            lo + hi * F::from(1 << 16) - word
         };
 
         array::IntoIter::new([