+++ examples/sha256/table16/compression/subregion_digest.rs
@@ -1,7 +1,6 @@
 use super::super::{super::DIGEST_SIZE, BlockWord, RoundWordDense};
 use super::{compression_util::*, CompressionConfig, State};
 use halo2::{
-    arithmetic::FieldExt,
     circuit::Region,
     pasta::pallas,
     plonk::{Advice, Column, Error},
@@ -41,7 +40,7 @@ impl CompressionConfig {
             a_5,
             abcd_row,
             || {
-                a.map(|a| pallas::Base::from_u64(a as u64))
+                a.map(|a| pallas::Base::from(a as u64))
                     .ok_or(Error::Synthesis)
             },
         )?;
@@ -63,7 +62,7 @@ impl CompressionConfig {
             a_5,
             efgh_row,
             || {
-                e.map(|e| pallas::Base::from_u64(e as u64))
+                e.map(|e| pallas::Base::from(e as u64))
                     .ok_or(Error::Synthesis)
             },
         )?;
@@ -102,7 +101,7 @@ impl CompressionConfig {
             word_col,
             row,
             || {
-                val.map(|val| pallas::Base::from_u64(val as u64))
+                val.map(|val| pallas::Base::from(val as u64))
                     .ok_or(Error::Synthesis)
             },
         )?;