+++ examples/sha256/table16/message_schedule/schedule_gates.rs
@@ -24,8 +24,8 @@ impl<F: FieldExt> ScheduleGate<F> {
         let hi = sigma_0_hi + sigma_1_hi + w_minus_9_hi + w_minus_16_hi;
 
         let word_check = lo
-            + hi * F::from_u64(1 << 16)
-            + (carry.clone() * F::from_u64(1 << 32) * (-F::one()))
+            + hi * F::from(1 << 16)
+            + (carry.clone() * F::from(1 << 32) * (-F::one()))
             + (word * (-F::one()));
         let carry_check = Gate::range_check(carry, 0, 3);
 
@@ -40,7 +40,7 @@ impl<F: FieldExt> ScheduleGate<F> {
         hi: Expression<F>,
         word: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
-        let check = lo + hi * F::from_u64(1 << 16) - word;
+        let check = lo + hi * F::from(1 << 16) - word;
         std::iter::empty().chain(Some(("s_decompose_0", s_decompose_0 * check)))
     }
 
@@ -58,9 +58,9 @@ impl<F: FieldExt> ScheduleGate<F> {
         word: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let decompose_check = a
-            + b * F::from_u64(1 << 3)
-            + c * F::from_u64(1 << 7)
-            + d * F::from_u64(1 << 18)
+            + b * F::from(1 << 3)
+            + c * F::from(1 << 7)
+            + d * F::from(1 << 18)
             + word * (-F::one());
         let range_check_tag_c = Gate::range_check(tag_c, 0, 2);
         let range_check_tag_d = Gate::range_check(tag_d, 0, 4);
@@ -91,12 +91,12 @@ impl<F: FieldExt> ScheduleGate<F> {
         word: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let decompose_check = a
-            + b * F::from_u64(1 << 3)
-            + c * F::from_u64(1 << 7)
-            + d * F::from_u64(1 << 10)
-            + e * F::from_u64(1 << 17)
-            + f * F::from_u64(1 << 18)
-            + g * F::from_u64(1 << 19)
+            + b * F::from(1 << 3)
+            + c * F::from(1 << 7)
+            + d * F::from(1 << 10)
+            + e * F::from(1 << 17)
+            + f * F::from(1 << 18)
+            + g * F::from(1 << 19)
             + word * (-F::one());
         let range_check_tag_d = Gate::range_check(tag_d, 0, 0);
         let range_check_tag_g = Gate::range_check(tag_g, 0, 3);
@@ -123,9 +123,9 @@ impl<F: FieldExt> ScheduleGate<F> {
         word: Expression<F>,
     ) -> impl Iterator<Item = (&'static str, Expression<F>)> {
         let decompose_check = a
-            + b * F::from_u64(1 << 10)
-            + c * F::from_u64(1 << 17)
-            + d * F::from_u64(1 << 19)
+            + b * F::from(1 << 10)
+            + c * F::from(1 << 17)
+            + d * F::from(1 << 19)
             + word * (-F::one());
         let range_check_tag_a = Gate::range_check(tag_a, 0, 1);
         let range_check_tag_d = Gate::range_check(tag_d, 0, 3);
@@ -140,7 +140,7 @@ impl<F: FieldExt> ScheduleGate<F> {
 
     /// b_lo + 2^2 * b_mid = b, on W_[1..49]
     fn check_b(b: Expression<F>, b_lo: Expression<F>, b_hi: Expression<F>) -> Expression<F> {
-        let expected_b = b_lo + b_hi * F::from_u64(1 << 2);
+        let expected_b = b_lo + b_hi * F::from(1 << 2);
         expected_b - b
     }
 
@@ -172,22 +172,22 @@ impl<F: FieldExt> ScheduleGate<F> {
                 .chain(Gate::three_bit_spread_and_range(a, spread_a.clone()));
         let check_b = Self::check_b(b, b_lo, b_hi);
         let spread_witness = spread_r0_even
-            + spread_r0_odd * F::from_u64(2)
-            + (spread_r1_even + spread_r1_odd * F::from_u64(2)) * F::from_u64(1 << 32);
+            + spread_r0_odd * F::from(2)
+            + (spread_r1_even + spread_r1_odd * F::from(2)) * F::from(1 << 32);
         let xor_0 = spread_b_lo.clone()
-            + spread_b_hi.clone() * F::from_u64(1 << 4)
-            + spread_c.clone() * F::from_u64(1 << 8)
-            + spread_d.clone() * F::from_u64(1 << 30);
+            + spread_b_hi.clone() * F::from(1 << 4)
+            + spread_c.clone() * F::from(1 << 8)
+            + spread_d.clone() * F::from(1 << 30);
         let xor_1 = spread_c.clone()
-            + spread_d.clone() * F::from_u64(1 << 22)
-            + spread_a.clone() * F::from_u64(1 << 50)
-            + spread_b_lo.clone() * F::from_u64(1 << 56)
-            + spread_b_hi.clone() * F::from_u64(1 << 60);
+            + spread_d.clone() * F::from(1 << 22)
+            + spread_a.clone() * F::from(1 << 50)
+            + spread_b_lo.clone() * F::from(1 << 56)
+            + spread_b_hi.clone() * F::from(1 << 60);
         let xor_2 = spread_d
-            + spread_a * F::from_u64(1 << 28)
-            + spread_b_lo * F::from_u64(1 << 34)
-            + spread_b_hi * F::from_u64(1 << 38)
-            + spread_c * F::from_u64(1 << 42);
+            + spread_a * F::from(1 << 28)
+            + spread_b_lo * F::from(1 << 34)
+            + spread_b_hi * F::from(1 << 38)
+            + spread_c * F::from(1 << 42);
         let xor = xor_0 + xor_1 + xor_2;
 
         check_spread_and_range
@@ -230,29 +230,29 @@ impl<F: FieldExt> ScheduleGate<F> {
                 ));
         // b_lo + 2^2 * b_mid + 2^4 * b_hi = b, on W_[49..62]
         let check_b1 = {
-            let expected_b = b_lo + b_mid * F::from_u64(1 << 2) + b_hi * F::from_u64(1 << 4);
+            let expected_b = b_lo + b_mid * F::from(1 << 2) + b_hi * F::from(1 << 4);
             expected_b - b
         };
         let spread_witness = spread_r0_even
-            + spread_r0_odd * F::from_u64(2)
-            + (spread_r1_even + spread_r1_odd * F::from_u64(2)) * F::from_u64(1 << 32);
+            + spread_r0_odd * F::from(2)
+            + (spread_r1_even + spread_r1_odd * F::from(2)) * F::from(1 << 32);
         let xor_0 = spread_b_lo.clone()
-            + spread_b_mid.clone() * F::from_u64(1 << 4)
-            + spread_b_hi.clone() * F::from_u64(1 << 8)
-            + spread_c.clone() * F::from_u64(1 << 14)
-            + spread_d.clone() * F::from_u64(1 << 18);
+            + spread_b_mid.clone() * F::from(1 << 4)
+            + spread_b_hi.clone() * F::from(1 << 8)
+            + spread_c.clone() * F::from(1 << 14)
+            + spread_d.clone() * F::from(1 << 18);
         let xor_1 = spread_c.clone()
-            + spread_d.clone() * F::from_u64(1 << 4)
-            + spread_a.clone() * F::from_u64(1 << 30)
-            + spread_b_lo.clone() * F::from_u64(1 << 50)
-            + spread_b_mid.clone() * F::from_u64(1 << 54)
-            + spread_b_hi.clone() * F::from_u64(1 << 58);
+            + spread_d.clone() * F::from(1 << 4)
+            + spread_a.clone() * F::from(1 << 30)
+            + spread_b_lo.clone() * F::from(1 << 50)
+            + spread_b_mid.clone() * F::from(1 << 54)
+            + spread_b_hi.clone() * F::from(1 << 58);
         let xor_2 = spread_d
-            + spread_a * F::from_u64(1 << 26)
-            + spread_b_lo * F::from_u64(1 << 46)
-            + spread_b_mid * F::from_u64(1 << 50)
-            + spread_b_hi * F::from_u64(1 << 54)
-            + spread_c * F::from_u64(1 << 60);
+            + spread_a * F::from(1 << 26)
+            + spread_b_lo * F::from(1 << 46)
+            + spread_b_mid * F::from(1 << 50)
+            + spread_b_hi * F::from(1 << 54)
+            + spread_c * F::from(1 << 60);
         let xor = xor_0 + xor_1 + xor_2;
 
         check_spread_and_range
@@ -294,31 +294,31 @@ impl<F: FieldExt> ScheduleGate<F> {
                 .chain(Gate::three_bit_spread_and_range(c, spread_c.clone()));
         let check_b = Self::check_b(b, b_lo, b_hi);
         let spread_witness = spread_r0_even
-            + spread_r0_odd * F::from_u64(2)
-            + (spread_r1_even + spread_r1_odd * F::from_u64(2)) * F::from_u64(1 << 32);
+            + spread_r0_odd * F::from(2)
+            + (spread_r1_even + spread_r1_odd * F::from(2)) * F::from(1 << 32);
         let xor_0 = spread_b_lo.clone()
-            + spread_b_hi.clone() * F::from_u64(1 << 4)
-            + spread_c.clone() * F::from_u64(1 << 8)
-            + spread_d.clone() * F::from_u64(1 << 14)
-            + spread_e.clone() * F::from_u64(1 << 28)
-            + spread_f.clone() * F::from_u64(1 << 30)
-            + spread_g.clone() * F::from_u64(1 << 32);
+            + spread_b_hi.clone() * F::from(1 << 4)
+            + spread_c.clone() * F::from(1 << 8)
+            + spread_d.clone() * F::from(1 << 14)
+            + spread_e.clone() * F::from(1 << 28)
+            + spread_f.clone() * F::from(1 << 30)
+            + spread_g.clone() * F::from(1 << 32);
         let xor_1 = spread_c.clone()
-            + spread_d.clone() * F::from_u64(1 << 6)
-            + spread_e.clone() * F::from_u64(1 << 20)
-            + spread_f.clone() * F::from_u64(1 << 22)
-            + spread_g.clone() * F::from_u64(1 << 24)
-            + spread_a.clone() * F::from_u64(1 << 50)
-            + spread_b_lo.clone() * F::from_u64(1 << 56)
-            + spread_b_hi.clone() * F::from_u64(1 << 60);
+            + spread_d.clone() * F::from(1 << 6)
+            + spread_e.clone() * F::from(1 << 20)
+            + spread_f.clone() * F::from(1 << 22)
+            + spread_g.clone() * F::from(1 << 24)
+            + spread_a.clone() * F::from(1 << 50)
+            + spread_b_lo.clone() * F::from(1 << 56)
+            + spread_b_hi.clone() * F::from(1 << 60);
         let xor_2 = spread_f
-            + spread_g * F::from_u64(1 << 2)
-            + spread_a * F::from_u64(1 << 28)
-            + spread_b_lo * F::from_u64(1 << 34)
-            + spread_b_hi * F::from_u64(1 << 38)
-            + spread_c * F::from_u64(1 << 42)
-            + spread_d * F::from_u64(1 << 48)
-            + spread_e * F::from_u64(1 << 62);
+            + spread_g * F::from(1 << 2)
+            + spread_a * F::from(1 << 28)
+            + spread_b_lo * F::from(1 << 34)
+            + spread_b_hi * F::from(1 << 38)
+            + spread_c * F::from(1 << 42)
+            + spread_d * F::from(1 << 48)
+            + spread_e * F::from(1 << 62);
         let xor = xor_0 + xor_1 + xor_2;
 
         check_spread_and_range
@@ -360,28 +360,28 @@ impl<F: FieldExt> ScheduleGate<F> {
                 .chain(Gate::three_bit_spread_and_range(c, spread_c.clone()));
         let check_b = Self::check_b(b, b_lo, b_hi);
         let spread_witness = spread_r0_even
-            + spread_r0_odd * F::from_u64(2)
-            + (spread_r1_even + spread_r1_odd * F::from_u64(2)) * F::from_u64(1 << 32);
+            + spread_r0_odd * F::from(2)
+            + (spread_r1_even + spread_r1_odd * F::from(2)) * F::from(1 << 32);
         let xor_0 = spread_d.clone()
-            + spread_e.clone() * F::from_u64(1 << 14)
-            + spread_f.clone() * F::from_u64(1 << 16)
-            + spread_g.clone() * F::from_u64(1 << 18);
+            + spread_e.clone() * F::from(1 << 14)
+            + spread_f.clone() * F::from(1 << 16)
+            + spread_g.clone() * F::from(1 << 18);
         let xor_1 = spread_e.clone()
-            + spread_f.clone() * F::from_u64(1 << 2)
-            + spread_g.clone() * F::from_u64(1 << 4)
-            + spread_a.clone() * F::from_u64(1 << 30)
-            + spread_b_lo.clone() * F::from_u64(1 << 36)
-            + spread_b_hi.clone() * F::from_u64(1 << 40)
-            + spread_c.clone() * F::from_u64(1 << 44)
-            + spread_d.clone() * F::from_u64(1 << 50);
+            + spread_f.clone() * F::from(1 << 2)
+            + spread_g.clone() * F::from(1 << 4)
+            + spread_a.clone() * F::from(1 << 30)
+            + spread_b_lo.clone() * F::from(1 << 36)
+            + spread_b_hi.clone() * F::from(1 << 40)
+            + spread_c.clone() * F::from(1 << 44)
+            + spread_d.clone() * F::from(1 << 50);
         let xor_2 = spread_g
-            + spread_a * F::from_u64(1 << 26)
-            + spread_b_lo * F::from_u64(1 << 32)
-            + spread_b_hi * F::from_u64(1 << 36)
-            + spread_c * F::from_u64(1 << 40)
-            + spread_d * F::from_u64(1 << 46)
-            + spread_e * F::from_u64(1 << 60)
-            + spread_f * F::from_u64(1 << 62);
+            + spread_a * F::from(1 << 26)
+            + spread_b_lo * F::from(1 << 32)
+            + spread_b_hi * F::from(1 << 36)
+            + spread_c * F::from(1 << 40)
+            + spread_d * F::from(1 << 46)
+            + spread_e * F::from(1 << 60)
+            + spread_f * F::from(1 << 62);
         let xor = xor_0 + xor_1 + xor_2;
 
         check_spread_and_range