+++ src/poly/multiopen.rs
@@ -250,6 +250,7 @@ where
 #[test]
 fn test_roundtrip() {
     use group::Curve;
+    use rand::rngs::OsRng;
 
     use super::commitment::{Blind, Params};
     use crate::arithmetic::{eval_polynomial, FieldExt};
@@ -260,6 +261,7 @@ fn test_roundtrip() {
 
     let params: Params<EqAffine> = Params::new(K);
     let domain = EvaluationDomain::new(1, K);
+    let rng = OsRng;
 
     let mut ax = domain.empty_coeff();
     for (i, a) in ax.iter_mut().enumerate() {
@@ -276,14 +278,14 @@ fn test_roundtrip() {
         *a = Fp::from(100 + i as u64);
     }
 
-    let blind = Blind(Fp::rand());
+    let blind = Blind(Fp::random(rng));
 
     let a = params.commit(&ax, blind).to_affine();
     let b = params.commit(&bx, blind).to_affine();
     let c = params.commit(&cx, blind).to_affine();
 
-    let x = Fp::rand();
-    let y = Fp::rand();
+    let x = Fp::random(rng);
+    let y = Fp::random(rng);
     let avx = eval_polynomial(&ax, x);
     let bvx = eval_polynomial(&bx, x);
     let cvy = eval_polynomial(&cx, y);
@@ -291,6 +293,7 @@ fn test_roundtrip() {
     let mut transcript = crate::transcript::Blake2bWrite::<_, _, Challenge255<_>>::init(vec![]);
     create_proof(
         &params,
+        rng,
         &mut transcript,
         std::iter::empty()
             .chain(Some(ProverQuery {